refactor: colocate CallInfo and EndReason with ConversationState
These types only exist as associated values of ConversationState; keep them in the same file and drop the Models junk-drawer entries.
