+++ Sources/ElevenLabs/Public/Conversation/ConversationState.swift
@@ -32,3 +32,13 @@ public enum ConversationState: Equatable, Sendable {
         return nil
     }
 }
+
+public struct CallInfo: Equatable, Sendable {
+    public let agentId: String
+    public let conversationId: String
+}
+
+public enum EndReason: Equatable, Sendable {
+    case userEnded
+    case remoteDisconnected
+}
