--- Sources/ElevenLabs/Public/Conversation/Models/CallInfo.swift
@@ -1,6 +1,0 @@
-import Foundation
-
-public struct CallInfo: Equatable, Sendable {
-    public let agentId: String
-    public let conversationId: String
-}
