--- Sources/ElevenLabs/Public/Conversation/Models/EndReason.swift
@@ -1,6 +1,0 @@
-import Foundation
-
-public enum EndReason: Equatable, Sendable {
-    case userEnded
-    case remoteDisconnected
-}
